refactor: move KvKey to separate module
The denokv.kv module is getting too big to easily manage.
