+++ src/denokv/__init__.py
@@ -14,9 +14,9 @@ from denokv.kv import CursorFormatType as CursorFormatType
 from denokv.kv import Kv as Kv
 from denokv.kv import KvCredentials as KvCredentials
 from denokv.kv import KvEntry as KvEntry
-from denokv.kv import KvKey as KvKey
 from denokv.kv import KvListOptions as KvListOptions
 from denokv.kv import KvU64 as KvU64
 from denokv.kv import ListKvEntry as ListKvEntry
 from denokv.kv import VersionStamp as VersionStamp
 from denokv.kv import open_kv as open_kv
+from denokv.kv_keys import KvKey as KvKey
